Refactoring DTOs
