+++ src/test/java/com/rapid7/client/dcerpc/msrrp/dto/Test_RegistryValueType.java
@@ -1,26 +1,30 @@
-/**
+/*
  * Copyright 2017, Rapid7, Inc.
  *
  * License: BSD-3-clause
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * * Redistributions of source code must retain the above copyright notice,
+ *   Redistributions of source code must retain the above copyright notice,
  * this list of conditions and the following disclaimer.
  *
- * * Redistributions in binary form must reproduce the above copyright
+ *  Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
  * documentation and/or other materials provided with the distribution.
  *
- * * Neither the name of the copyright holder nor the names of its contributors
+ *  Neither the name of the copyright holder nor the names of its contributors
  * may be used to endorse or promote products derived from this software
  * without specific prior written permission.
+ *
+ *
  */
-package com.rapid7.client.dcerpc.msrrp;
+package com.rapid7.client.dcerpc.msrrp.dto;
 
 import org.junit.Test;
 
-import static com.rapid7.client.dcerpc.msrrp.RegistryValueType.*;
+import com.rapid7.client.dcerpc.msrrp.dto.RegistryValueType;
+
+import static com.rapid7.client.dcerpc.msrrp.dto.RegistryValueType.*;
 import static org.junit.Assert.*;
 
 public class Test_RegistryValueType {
