+++ src/test/java/com/rapid7/client/dcerpc/msrrp/messages/Test_BaseRegEnumValueResponse.java
@@ -20,7 +20,7 @@ package com.rapid7.client.dcerpc.msrrp.messages;
 
 import java.io.IOException;
 import org.junit.Test;
-import com.rapid7.client.dcerpc.msrrp.RegistryValueType;
+import com.rapid7.client.dcerpc.msrrp.dto.RegistryValueType;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
@@ -93,7 +93,7 @@ public class Test_BaseRegEnumValueResponse {
         response.fromHexString("1600feff00000200ff7f0000000000000b000000530079007300740065006d0052006f006f0074000000000004000200010000000800020016000000000000001600000043003a005c00570069006e0064006f0077007300000000000c00020016000000100002001600000000000000");
 
         assertEquals("SystemRoot", response.getName().getValue());
-        assertEquals(RegistryValueType.REG_SZ, response.getType());
+        assertEquals(RegistryValueType.REG_SZ.getTypeID(), (int) response.getType());
         assertArrayEquals("C:\\Windows\0".getBytes("UTF-16LE"), response.getData().getArray());
         assertEquals(0, response.getReturnValue());
     }
